requirements.txt
pyyaml
pyinotify
